[Updates and Notes] Move to independent page; Link from Homepage
