+++ src/common/Header.css
@@ -142,7 +142,7 @@
   }
 }
 
-@media screen and (max-width: 1023px) {
+@media screen and (max-width: 1065px) {
   .nav-container {
     padding-left: 1rem;
     padding-right: 1rem;
