+++ src/data-publication/index.js
@@ -15,7 +15,6 @@ import {
 import DynamicDataset from './reports/DynamicDataset'
 import NotFound from '../common/NotFound'
 import { withAppContext } from '../common/appContextHOC.jsx'
-import PublicationChanges from './ChangeLog/'
 
 import './index.css'
 
@@ -29,11 +28,6 @@ const DataPublication = ({ config }) => {
           exact
           from='/data-publication'
           to={`/data-publication/${shared[0]}`}
-        />
-        <Route
-          exact
-          path='/data-publication/updates'
-          component={PublicationChanges}
         />
         <Route path='/data-publication/documents' component={SupportingDocs} />
         <Route
