+++ src/homepage/ForDataUsers/ChangeLog.jsx
@@ -3,7 +3,7 @@ import { ExpandableCard } from '../ExpandableCard'
 export const ChangeLog = () => (
   <ExpandableCard
     title='HMDA Updates and Notes'
-    destination='/data-publication/updates'
+    destination='/updates-notes'
     description="Tracking releases, updates, and corrections to HMDA's publications, data products, documentation, and tools."
     disableExpansion={true}
   />
